+++ apps/api-harmonization/src/blocks/article-details/article-details.service.ts
@@ -22,7 +22,7 @@ export class ArticleDetailsService {
         headers: AppHeaders,
     ): Observable<ArticleDetailsBlock> {
         const cms = this.cmsService.getArticleDetailsBlock({ ...query, locale: headers['x-locale'] });
-        const article = this.articleService.getArticle(params);
+        const article = this.articleService.getArticle({ ...params, locale: headers['x-locale'] });
 
         return forkJoin([article, cms]).pipe(
             map(([article, cms]) => {