+++ apps/api-harmonization/src/blocks/article-list/article-list.controller.ts
@@ -1,10 +1,10 @@
-import { Controller, Get, Headers, Query, UseInterceptors } from '@nestjs/common';
+import { Body, Controller, Get, Headers, Query, UseInterceptors } from '@nestjs/common';
 import { LoggerService } from '@o2s/utils.logger';
 
 import { AppHeaders } from '@o2s/api-harmonization/utils/headers';
 
 import { URL } from './';
-import { GetArticleListBlockQuery } from './article-list.request';
+import { GetArticleListBlockBody, GetArticleListBlockQuery } from './article-list.request';
 import { ArticleListService } from './article-list.service';
 
 @Controller(URL)
@@ -13,7 +13,11 @@ export class ArticleListController {
     constructor(protected readonly service: ArticleListService) {}
 
     @Get()
-    getArticleListBlock(@Headers() headers: AppHeaders, @Query() query: GetArticleListBlockQuery) {
-        return this.service.getArticleListBlock(query, headers);
+    getArticleListBlock(
+        @Headers() headers: AppHeaders,
+        @Query() query: GetArticleListBlockQuery,
+        @Body() body: GetArticleListBlockBody,
+    ) {
+        return this.service.getArticleListBlock(query, headers, body);
     }
 }