+++ apps/api-harmonization/src/blocks/article-list/article-list.module.ts
@@ -1,7 +1,7 @@
 import { DynamicModule, Module } from '@nestjs/common';
 import { Type } from '@nestjs/common/interfaces/type.interface';
 
-import { ApiConfig, Articles, CMS } from '@o2s/framework/modules';
+import { ApiConfig, Articles, CMS, Search } from '@o2s/framework/modules';
 
 import { ArticleListController } from './article-list.controller';
 import { ArticleListService } from './article-list.service';
@@ -11,6 +11,7 @@ export class ArticleListBlockModule {
     static register(config: ApiConfig): DynamicModule {
         const cmsService = config.integrations.cms.service;
         const articleService = config.integrations.articles.service;
+        const searchService = config.integrations.search.service;
 
         return {
             module: ArticleListBlockModule,
@@ -23,6 +24,10 @@ export class ArticleListBlockModule {
                 {
                     provide: Articles.Service,
                     useClass: articleService as Type,
+                },
+                {
+                    provide: Search.Service,
+                    useClass: searchService as Type,
                 },
             ],
             controllers: [ArticleListController],