+++ apps/api-harmonization/src/blocks/article-list/article-list.service.ts
@@ -7,7 +7,7 @@ import { AppHeaders } from '@o2s/api-harmonization/utils/headers';
 
 import { mapArticleList } from './article-list.mapper';
 import { ArticleListBlock } from './article-list.model';
-import { GetArticleListBlockQuery } from './article-list.request';
+import { GetArticleListBlockBody, GetArticleListBlockQuery } from './article-list.request';
 
 @Injectable()
 export class ArticleListService {
@@ -16,9 +16,13 @@ export class ArticleListService {
         private readonly articleService: Articles.Service,
     ) {}
 
-    getArticleListBlock(query: GetArticleListBlockQuery, headers: AppHeaders): Observable<ArticleListBlock> {
+    getArticleListBlock(
+        query: GetArticleListBlockQuery,
+        headers: AppHeaders,
+        body: GetArticleListBlockBody,
+    ): Observable<ArticleListBlock> {
         const cms = this.cmsService.getArticleListBlock({ ...query, locale: headers['x-locale'] });
-        const articles = this.articleService.getArticleList(query);
+        const articles = this.articleService.getArticleList({ ...query, locale: headers['x-locale'] }, body);
 
         return forkJoin([articles, cms]).pipe(
             map(([articles, cms]) => mapArticleList(articles, cms, headers['x-locale'])),