+++ packages/framework/src/modules/articles/index.ts
@@ -1,5 +1,5 @@
 export * as Model from './articles.model';
 export * as Request from './articles.request';
-export { ArticleService as Service } from './articles.service';
+export { ArticlesService as Service } from './articles.service';
 export { ArticleController as Controller } from './articles.controller';
-export { ArticleModule as Module } from './articles.module';
+export { ArticlesModule as Module } from './articles.module';