+++ packages/framework/src/modules/billing-accounts/billing-accounts.module.ts
@@ -1,11 +1,12 @@
 import { HttpModule } from '@nestjs/axios';
-import { DynamicModule, Module } from '@nestjs/common';
+import { DynamicModule, Global, Module } from '@nestjs/common';
 import { Type } from '@nestjs/common/interfaces/type.interface';
 
 import { BillingAccountController } from './billing-accounts.controller';
 import { BillingAccountService } from './billing-accounts.service';
 import { ApiConfig } from '@/api-config';
 
+@Global()
 @Module({})
 export class BillingAccountModule {
     static register(config: ApiConfig): DynamicModule {