+++ apps/api-harmonization/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @o2s/api-harmonization
 
+## 0.14.0
+
+### Minor Changes
+
+- 80b678a: Added search integration with Algolia
+
+### Patch Changes
+
+- Updated dependencies [80b678a]
+    - @o2s/integrations.strapi-cms@0.13.0
+    - @o2s/integrations.algolia@0.2.0
+    - @o2s/integrations.mocked@0.13.0
+    - @o2s/framework@0.14.0
+
 ## 0.13.0
 
 ### Minor Changes